feat: 더보기페이지 통합 및 프로젝트 get api 연결
