--- src/pages/RightPages/MyPage/MorePage/AlgorithmMorePage.module.css
@@ -1,51 +1,0 @@
-.algorithmSection{
-    padding: 1rem 2rem;
-
-    container-type: inline-size; 
-    container-name: algorithm_section;
-}
-
-.algorithmMore{
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 0.5rem;
-    
-    color: var(--color-text-primary);
-    font-size: var(--text-base);
-    font-weight: var(--font-weight-semibold);
-    font-family: var(--font-general);
-}
-
-.no_data {
-    width: 100%;
-    padding: 3rem;
-    text-align: center;
-    line-height: 1.5rem;
-    color: var(--color-text-sub);
-    font-size: var(--text-lg);
-}
-
-::-webkit-scrollbar {
-    display: none;
-}
-
-
-.algorithmList{
-    display: grid;
-	grid-template-columns: repeat(1, 1fr);
-    gap: 1rem;
-    padding: 1rem;
-}
-
-@container algorithm_section (min-width: 600px) {
-    .algorithmList {
-        grid-template-columns: repeat(2, 1fr);
-    }
-}
-
-@container algorithm_section (min-width: 900px) {
-    .algorithmList {
-        grid-template-columns: repeat(3, 1fr);
-    }
-}
